requirements.txt
requests~=2.30.0
python-dotenv~=1.0.1
pip~=22.2.2
attrs~=23.2.0
wheel~=0.43.0
pillow~=10.2.0
tornado~=6.3.1
Jinja2~=3.1.3
click~=8.1.7
ipython~=8.13.0
certifi~=2023.5.7
idna~=3.4
urllib3~=2.0.2
setuptools~=63.2.0
filelock~=3.14.0
packaging~=23.1
config~=0.5.1
platformdirs~=3.5.0